+++ pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   intl: ^0.19.0
   scroll_to_index: ^3.0.1
   cached_network_image: ^3.3.1
+  rxdart: ^0.27.7
 dev_dependencies:
   flutter_test:
     sdk: flutter
